# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie API",
        "description": "API for managing films and screening",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/vetitesek": {
            "get": {
                "tags": [
                    "Vetitesek"
                ],
                "summary": "Get list of all screenings",
                "operationId": "49bc639584d68925757f772a5be41565",
                "responses": {
                    "200": {
                        "description": "List of all screenings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vetitesek"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/vetitesek/create": {
            "post": {
                "tags": [
                    "Vetitesek"
                ],
                "summary": "Create a new Vetitesek record",
                "description": "This endpoint allows you to create a new Vetitesek record with time, available seats, and film ID.",
                "operationId": "21664f3d276897b44abe4923a3c32e1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "time",
                                    "available_seats",
                                    "film_id"
                                ],
                                "properties": {
                                    "time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-03-27 19:00:00"
                                    },
                                    "available_seats": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "film_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vetitesek successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27 19:00:00"
                                        },
                                        "available_seats": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "film_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27T10:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27T10:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vetitesek/show/{id}": {
            "get": {
                "tags": [
                    "Vetitesek"
                ],
                "summary": "Get a single Vetitesek record by ID",
                "description": "This endpoint allows you to fetch a Vetitesek record based on the provided ID.",
                "operationId": "331eb7d9c31886033a519aa66e371729",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Vetitesek record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the Vetitesek record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27 19:00:00"
                                        },
                                        "available_seats": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "film_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27T10:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27T10:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vetitesek record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vetitesek not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vetitesek/update/{id}": {
            "put": {
                "tags": [
                    "Vetitesek"
                ],
                "summary": "Update a Vetitesek (Screening) record",
                "description": "This endpoint allows you to update an existing Vetitesek (screening) record with new time, available seats, and film ID.",
                "operationId": "2f00921769ef826bd387c9b3070f54a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Vetitesek (screening) to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "time",
                                    "available_seats",
                                    "film_id"
                                ],
                                "properties": {
                                    "time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-03-27 19:00:00"
                                    },
                                    "available_seats": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "film_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vetitesek successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27 19:00:00"
                                        },
                                        "available_seats": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "film_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27T10:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-27T10:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Screening not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Screening not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vetitesek/delete/{id}": {
            "delete": {
                "tags": [
                    "Vetitesek"
                ],
                "summary": "Delete a Vetitesek (Screening) record",
                "description": "This endpoint allows you to delete a Vetitesek (screening) record by its ID.",
                "operationId": "62e33993fa939f8a7ae90dd64dac1740",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Vetitesek (screening) to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Screening successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Screening deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Screening not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Screening not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid ID format"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Film": {
                "required": [
                    "id",
                    "title",
                    "description",
                    "age_rating",
                    "language",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Matrix"
                    },
                    "description": {
                        "type": "string",
                        "example": "A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers."
                    },
                    "age_rating": {
                        "type": "integer",
                        "example": 18
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "cover_image": {
                        "type": "string",
                        "example": "matrix_cover.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-26T18:56:49"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-26T18:56:49"
                    }
                },
                "type": "object"
            },
            "Vetitesek": {
                "required": [
                    "time",
                    "available_seats",
                    "film_id"
                ],
                "properties": {
                    "time": {
                        "description": "The time when the film is scheduled to be shown",
                        "type": "string",
                        "format": "date-time"
                    },
                    "available_seats": {
                        "description": "The number of available seats for the screening",
                        "type": "integer"
                    },
                    "film_id": {
                        "description": "The ID of the film being screened",
                        "type": "integer"
                    },
                    "film": {
                        "$ref": "#/components/schemas/Film"
                    }
                },
                "type": "object"
            }
        }
    },
...
}